Further work on openSSL support.  OpenSSL 1.0 to be used.

+++ source/mailbox.d
@@ -44,7 +44,7 @@ public:
     return m_protocol;
   }
 
-  @property int size()
+  @property size_t size()
   {
     return m_connection.m_messages.length;
   }
